requirements.txt
google-cloud-dataplex
google-cloud-storage
google-cloud-bigquery
PyYAML

